fix: mounts not propaged to storage-csi node.
when a new zfs dataset was created and mounted on the host after the pod
had already started it would not be mounted inside the pod and would
cause it to use the wrong directory when mounting a persistent volume
claim. this would lead to dataloss after the storage-csi pod was removed
for whatever reason.