dataset reuse, dataset name change
datasets are now named as <parent dataset>/<namespace>-<pvc> so that
it is easier to know what dataset corresponds to what pvc.
when a dataset is deleted a timestamp is append at the end.
the properties remain unchanged.

datasets are now reused if possible. if a pvc is created but a dataset
for that namespace/pvc already exists and its deleted property is false
then that dataset will be used instead of creating a new one.
this should allow for the cluster to be redeployed from scratch.